+++ scirs2-stats/src/distributions/negative_binomial.rs
@@ -5,7 +5,6 @@
 use crate::error::{StatsError, StatsResult};
 use crate::sampling::SampleableDistribution;
 use num_traits::{Float, NumCast};
-use rand::rng;
 use rand_distr::Distribution;
 use scirs2_core::Rng;
 use statrs::function::gamma::ln_gamma;
@@ -378,7 +377,7 @@ impl<F: Float + NumCast + std::fmt::Display> NegativeBinomial<F> {
                     // Generate geometric random variable (# failures before first success)
                     let u: f64 = rng.gen_range(0.0..1.0);
                     let p_f64 = <f64 as num_traits::NumCast>::from(self.p).unwrap_or(0.5);
-                    let geom_sample = (u.ln() / (1.0 - p_f64).ln()).floor() as usize;
+                    let geom_sample = (u.ln() / ((1.0_f64 - p_f64) as f64).ln()).floor() as usize;
                     sum += geom_sample;
                 }
                 samples.push(F::from(sum).unwrap());
